Removed statement on certification
The statement:
```Apex.AI and ADLINK are working to make the Eclipse Cyclone DDS with iceoryx open source functionally safe and ISO 26262 certifiable. You can read about that [here](https://www.apex.ai/apex-middleware)``` 

Is not true as ADLINK is not working with APEX.AI on functionally safe version of Cyclone.